pyproject.toml
[project]
name = "evvo-works"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django>=5.2",
    "django-rest-framework>=0.1.0",
    "djangorestframework-simplejwt>=5.5.0",
    "django-cors-headers>=4.7.0",
]
